fixed layout approach for nav
