+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import './App.css';
-import Navbar from './components/layout';
+import Layout from './components/layout';
 import Categories from './components/categories';
 import Books from './components/showbooks';
 
@@ -9,7 +9,7 @@ function App() {
   return (
     <div className="wraper">
       <Routes>
-        <Route element={<Navbar />}>
+        <Route element={<Layout />}>
           <Route index element={<Books />} />
           <Route path="Categories" element={<Categories />} />
         </Route>
