+++ src/components/layout.css
@@ -1,6 +1,9 @@
 .navbar {
   width: 100%;
   position: absolute;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
 
 .navdiv {
@@ -65,6 +68,7 @@
 }
 
 .header {
+  padding-left: 1rem;
   font-size: 2rem;
   font-weight: 700;
 }
